backend/requirements.txt
# core
haystack-ai
fastapi
uvicorn[standard]
python-multipart
python-dotenv

# helpers
pypdf
sentence-transformers
openai
